pyproject.toml
[project]
name = "thesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ale-py>=0.10.1",
    "numpyro>=0.17.0",
    "opencv-python>=4.11.0.86",
    "pymc>=5.20.1",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
]
